bucket_storage_agent/requirements.txt
google-adk
python-dotenv
litellm
pydantic
google-cloud-storage
google-cloud-monitoring
google-cloud-bigquery
google-auth

# Web UI dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
websockets>=11.0
httpx>=0.27.0
